create contactForm and other changes
